Improve limit performance for table mode

+++ flink-libraries/flink-sql-client/src/main/java/org/apache/flink/table/client/config/Execution.java
@@ -92,8 +92,8 @@ public class Execution {
 		return Integer.parseInt(properties.getOrDefault(PropertyStrings.EXECUTION_MAX_PARALLELISM, Integer.toString(128)));
 	}
 
-	public long getMaxTableResultRows() {
-		return Long.parseLong(properties.getOrDefault(PropertyStrings.EXECUTION_MAX_TABLE_RESULT_ROWS, Long.toString(1_000_000)));
+	public int getMaxTableResultRows() {
+		return Integer.parseInt(properties.getOrDefault(PropertyStrings.EXECUTION_MAX_TABLE_RESULT_ROWS, Integer.toString(1_000_000)));
 	}
 
 	public RestartStrategies.RestartStrategyConfiguration getRestartStrategy() {
